provider/bluecat: minor test cleanup
* Move tests to appropriate package
* Add test for SplitProperties function
* Remove unused helper function
* Use table driven tests
